Added two new custom exceptions ClientException & RequestException to easier catch when request/client fail in apps using Buzz. Closes #79
