+++ lib/Buzz/Exception/ClientException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Buzz\Exception;
+
+class ClientException extends \RuntimeException
+{
+
+}
