+++ lib/Buzz/Exception/RequestException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Buzz\Exception;
+
+class RequestException extends \ErrorException
+{
+
+}
